fix return

+++ app/controllers/receptions_controller.rb
@@ -58,7 +58,8 @@ class ReceptionsController < ApplicationController
   def destroy
     reception_id = destroy_params
     reception = current_user.reception.find(reception_id)
-    render_400_if_reserved(reception) && return
+    return if render_400_if_reserved(reception)
+
     response = {}
     if reception.destroy!
       response['reception_id'] = reception.id
